refactor(board): 의존성 순환 해결을 위해 DTO를 App 계층으로 이동
- Service가 DTO를 사용하므로, DTO가 Presentation에 있으면 역방향 의존 발생
- 별도의 Service DTO를 만드는 대신, 실용성을 위해 Application 계층에 통합하여 관리
- ArchitectureTest 검사 범위를 board 패키지로 한정하여 우선 적용
