+++ backend/src/main/java/com/ssafy/dash/board/application/dto/BoardUpdateRequest.java
@@ -1,4 +1,4 @@
-package com.ssafy.dash.board.presentation.dto;
+package com.ssafy.dash.board.application.dto;
 
 public class BoardUpdateRequest {
